updated CSS file
update .score to show as a circle
Add a green background colorand white text to .likes button.
Add a light-green background and green text to .likes button on
mouseover/hover.
Add a red background color and white text to your .dislikes button.
Add a light-red background and red text to your .dislikes button on
mouseover/hover.
Add a smooth transition when hovering on and off .likes and .dislikes